+++ src/Resources/SystemPermissions.php
@@ -6,6 +6,8 @@ namespace Authn\Sdk\Resources;
 
 final class SystemPermissions
 {
+    public const ORG_SYS_PROFILE_READ = 'org:sys_profile:read';
+
     public const ORG_SYS_PROFILE_MANAGE = 'org:sys_profile:manage';
 
     public const ORG_SYS_PROFILE_DELETE = 'org:sys_profile:delete';
@@ -14,19 +16,19 @@ final class SystemPermissions
 
     public const ORG_SYS_MEMBERSHIPS_MANAGE = 'org:sys_memberships:manage';
 
+    public const ORG_SYS_INVITATIONS_READ = 'org:sys_invitations:read';
+
+    public const ORG_SYS_INVITATIONS_MANAGE = 'org:sys_invitations:manage';
+
     public const ORG_SYS_DOMAINS_READ = 'org:sys_domains:read';
 
     public const ORG_SYS_DOMAINS_MANAGE = 'org:sys_domains:manage';
 
-    public const ORG_SYS_BILLING_READ = 'org:sys_billing:read';
+    public const ORG_SYS_ROLES_READ = 'org:sys_roles:read';
 
-    public const ORG_SYS_BILLING_MANAGE = 'org:sys_billing:manage';
+    public const ORG_SYS_ROLES_MANAGE = 'org:sys_roles:manage';
 
-    public const ORG_SYS_SSO_READ = 'org:sys_sso:read';
+    public const ORG_SYS_PERMISSIONS_READ = 'org:sys_permissions:read';
 
-    public const ORG_SYS_SSO_MANAGE = 'org:sys_sso:manage';
-
-    public const ORG_SYS_PROVISIONING_READ = 'org:sys_provisioning:read';
-
-    public const ORG_SYS_PROVISIONING_MANAGE = 'org:sys_provisioning:manage';
+    public const ORG_SYS_PERMISSIONS_MANAGE = 'org:sys_permissions:manage';
 }